# api/openapi.yaml
openapi: 3.0.1
info:
  title: eMekdep API
  description: ''
  version: 1.0.0
...
paths:
...
  /users/notifications:
    get:
      summary: /users/notifications
      deprecated: false
      description: ''
      tags:
        - App
      parameters:
        - name: user_id
          in: query
          description: ''
          required: false
          schema:
            type: integer
        - name: id[]
          in: query
          description: ''
          required: false
          schema:
            type: integer
        - name: read
          in: query
          description: ''
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      total:
                        type: integer
                      total_unread:
                        type: integer
                      items:
                        type: array
                        items: &ref_6
                          $ref: '#/components/schemas/UserNotification'
                    type: object
      security: []
  /users/notifications/{id}:
    put:
      summary: /users/notifications/{id}
      deprecated: false
      description: ''
      tags:
        - App
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserNotificationCreate'
            example: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      item: *ref_6
                    type: object
      security: []
  /users/notifications/sender:
    get:
      summary: /users/notifications/sender
      deprecated: false
      description: ''
      tags:
        - App
      parameters:
        - name: user_id
          in: query
          description: ''
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      items:
                        type: array
                        items: &ref_7
                          $ref: '#/components/schemas/Notification'
                    type: object
      security: []
    post:
      summary: /users/notifications/sender
      deprecated: false
      description: ''
      tags:
        - App
      parameters: []
      requestBody:
        content:
          application/json:
            schema: &ref_8
              $ref: '#/components/schemas/NotificationCreate'
            example: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      item: *ref_7
                    type: object
      security: []
  /users/notifications/sender/{id}:
    put:
      summary: /users/notifications/sender/{id}
      deprecated: false
      description: ''
      tags:
        - App
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema: *ref_8
            example: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      item: *ref_7
                    type: object
      security: []
  /lesson/like/{id}:
    post:
      summary: /lesson/like/{id}
      deprecated: false
      description: ''
      tags:
        - Journal
      parameters:
        - name: id
          in: path
          description: ''
          required: true
          example: ''
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LessonLikesRequest'
            example: ''
      responses:
        '200':
          description: Successfully created a lesson like
          content:
            application/json:
              schema:
                type: object
                properties: {}
      security: []
  /lessons/journal:
    get:
      summary: /lessons/journal
      deprecated: false
      description: ''
      tags:
        - Journal
      parameters:
        - name: date
          in: query
          description: ''
          required: false
          schema:
            type: string
        - name: subject_id
          in: query
          description: ''
          required: true
          schema:
            type: string
        - name: period_number
          in: query
          description: ''
          required: false
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      data:
                        $ref: '#/components/schemas/Journal'
                    type: object
      security: []
  /lessons:
    post:
      summary: /lessons
      deprecated: false
      description: ''
      tags:
        - Journal
      parameters: []
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                title:
                  type: string
                  examples:
                    - Homework Big Text...
                  example: Homework Big Text...
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                files_delete:
                  type: array
                  items: &ref_38
                    $ref: '#/components/schemas/FileDelete'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      lesson: &ref_55
                        $ref: '#/components/schemas/Lesson'
                      grade: &ref_10
                        $ref: '#/components/schemas/Grade'
                      absent: &ref_56
                        $ref: '#/components/schemas/Absent'
                    type: object
      security: []
  /lessons/final:
    get:
      summary: /lessons/final
      deprecated: false
      description: ''
      tags:
        - Journal
      parameters:
        - name: subject_id
          in: query
          description: ''
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                properties:
                  students:
                    type: array
                    items:
                      type: object
                      properties:
                        student: *ref_9
                        periods:
                          type: object
                          properties:
                            '1': *ref_10
                        exam: *ref_10
                        final: *ref_10
                type: object
      security: []
    post:
      summary: /lessons/final
      deprecated: false
      description: ''
      tags:
        - Journal
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - &ref_37
                  $ref: '#/components/schemas/GradeCreate'
                - properties:
                    subject_id: *ref_0
                  type: object
            example: ''
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - *ref_5
                  - properties:
                      grade: *ref_10
                    type: object
      security: []
...
components:
  schemas:
...
    UserNotificationCreate:
      properties:
        comment:
          type: string
        comment_files:
          type: array
          items: &ref_36
            $ref: '#/components/schemas/File'
        comment_files_delete:
          type: array
          items: *ref_36
      type: object
...
    NotificationCreate:
      properties:
        school_ids:
          description: create_only
          type: array
          items:
            type: integer
        user_ids:
          description: create_only
          type: array
          items: *ref_0
        roles:
          type: array
          items:
            type: string
        title:
          type: string
        content:
          type: string
        files:
          type: array
          items: *ref_36
        files_delete:
          type: array
          items: *ref_36
      type: object
...
    GradeCreate:
      properties:
        student_id: *ref_0
        student_ids:
          type: array
          items: *ref_0
        value:
          type: integer
          examples:
            - 5
        values:
          type: array
          items:
            type: integer
          examples:
            - - 5
              - 2
        comment:
          type: string
      type: object
...
    UserNotification:
      properties:
        id: *ref_0
        notification: *ref_7
        user_id: *ref_0
        read_at: *ref_4
        comment:
          type: string
        comment_files:
          type: array
          items: *ref_36
      type: object
    Notification:
      properties:
        id: *ref_0
        school_ids:
          type: array
          items:
            type: integer
        user_ids:
          type: array
          items: *ref_0
        roles:
          type: array
          items:
            type: string
        author_id: *ref_0
        title:
          type: string
        content:
          type: string
        files:
          type: array
          items: *ref_36
        updated_at: *ref_4
        created_at: *ref_4
      type: object
...
    Journal:
      properties:
        students:
          type: array
          items: *ref_9
        period_grades:
          type: array
          items: *ref_52
        student_notes:
          type: array
          items:
            $ref: '#/components/schemas/StudentNote'
        lessons:
          type: array
          items:
            properties:
              lesson: *ref_55
              grades:
                type: array
                items: *ref_10
              absents:
                type: array
                items: *ref_56
            type: object
      type: object
    Assignment:
      properties:
        id: *ref_0
        content:
          type: string
          examples:
            - Homework Big Text...
        created_at: *ref_4
        updated_at: *ref_4
      type: object
    StudentNote:
      properties:
        id: *ref_0
        note:
          type: string
          examples:
            - qwertyui
        student_id: *ref_0
        created_at: *ref_4
        updated_at: *ref_4
      type: object
    Absent:
      properties:
        id: *ref_0
        reason:
          type: string
          examples:
            - qwertyui
        lesson_id:
          type: string
          examples:
            - qwertyui
        comment:
          type: string
          examples:
            - qwertyui
        created_at: *ref_4
        updated_at: *ref_4
      type: object
...
    Grade:
      properties:
        id: *ref_0
        value:
          type: integer
          examples:
            - 5
        values:
          type: array
          items:
            type: integer
          examples:
            - - 5
              - 2
        comment:
          type: string
          examples:
            - qwertyui
        lesson_id:
          type: string
          examples:
            - qwertyui
        created_at: *ref_4
        updated_at: *ref_4
      type: object
    Lesson:
      properties:
        id: *ref_0
        title:
          type: string
          examples:
            - Tema1
        content:
          type: string
          examples:
            - Tema Big Text...
        date:
          type: string
          examples:
            - '2023-10-18'
        hour_number:
          type: integer
          examples:
            - 1
        assignment:
          $ref: '#/components/schemas/Assignment'
      type: object
...
    LessonLikesRequest:
      type: object
      properties:
        id:
          type: integer
          format: uint
        user_id:
          type: integer
          format: uint
        lesson_id:
          type: integer
          format: uint
...
    FileDelete:
      description: File from storage
      type: string
      examples:
        - files/file.pdf
    File:
      description: File from storage
      type: string
      examples:
        - https://storage.emekdep.com/files/file.pdf
...
  securitySchemes: {}
servers: []